+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.2.0 (????-??-??)
+0.2.0 (2024-02-27)
 ------------------
 
 - added support for XTuner conversation JSON format: `from-xtuner` and `to-xtuner`